API-650: Split App in Read and Write models
