+++ src/Akeneo/Apps/back/Application/Command/CreateAppCommand.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace Akeneo\Apps\Application\Command;
 
-use Akeneo\Apps\Domain\Model\AppCode;
-use Akeneo\Apps\Domain\Model\AppLabel;
-use Akeneo\Apps\Domain\Model\FlowType;
+use Akeneo\Apps\Domain\Model\Write\AppCode;
+use Akeneo\Apps\Domain\Model\Write\AppLabel;
+use Akeneo\Apps\Domain\Model\Write\FlowType;
 
 /**
  * @author Romain Monceau <romain@akeneo.com>
