+++ src/Akeneo/Apps/back/Application/Command/CreateAppHandler.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Akeneo\Apps\Application\Command;
 
-use Akeneo\Apps\Domain\Model\App;
+use Akeneo\Apps\Domain\Model\Write\App;
 use Akeneo\Apps\Domain\Persistence\Repository\AppRepository;
 
 /**
